Software_Systems/Django_Application/Degree_Checklist_Project/DegreeFinal-main/requirements.txt
asgiref==3.7.2
contourpy==1.2.0
cycler==0.12.1
Django==4.2.5
django-crispy-forms==2.1
djangorestframework==3.14.0
fonttools==4.46.0
gunicorn==21.2.0
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
reportlab==4.0.7
six==1.16.0
sqlparse==0.4.4
whitenoise==6.6.0
